Fix LED output corruption isses
Some LEDs were not being updated/written to. This is due to some
interaction between receiving serial data and the Teensy FastLED
library. The teensy serial ports use interrupts and a 4-byte buffer, and
FastLED uses DMA with interrupts. This change fixes the issue by
enforcing timing in the brain and interface code. It has a strict
sequence, where the brain will receive the serial data at a specific
time (when it is not writing out LEDs).